+++ mapsforge_flutter/lib/src/rendertheme/rule/positiverule.dart
@@ -14,7 +14,7 @@ class PositiveRule extends Rule {
   PositiveRule(RuleBuilder ruleBuilder, this.keyMatcher, this.valueMatcher) : super(ruleBuilder);
 
   @override
-  bool matchesNode(List<Tag> tags, int zoomLevel, double indoorLevel) {
+  bool matchesNode(List<Tag> tags, int zoomLevel, int indoorLevel) {
     return this.zoomMin <= zoomLevel &&
         this.zoomMax >= zoomLevel &&
         IndoorNotationMatcher.isOutdoorOrMatchesIndoorLevel(tags, indoorLevel) &&
@@ -24,7 +24,7 @@ class PositiveRule extends Rule {
   }
 
   @override
-  bool matchesWay(List<Tag> tags, int zoomLevel, double indoorLevel, Closed closed) {
+  bool matchesWay(List<Tag> tags, int zoomLevel, int indoorLevel, Closed closed) {
     return this.zoomMin <= zoomLevel &&
         this.zoomMax >= zoomLevel &&
         IndoorNotationMatcher.isOutdoorOrMatchesIndoorLevel(tags, indoorLevel) &&
